Same logger, more versatile!
The leemons-logger has been upgraded, allowing the users to decide which
formatters uses each transport, and including a multi-thread logger.

The multi-thread logger manages multiple clusters, allowing only the
master thread to actualy output the log (to files and console) and the
workers are limited to emit the log to the master process. Allowing us
to handle everything from the master.

The logger now supports custom labels, allowing you to use them for
distinguishing between different services (or if the process was the
master or the worker).

Another important feature is the log to a file, which is no longer
prettified and is save in a json-like style, more commonly supported by third parties.